Refactor: navigation 타입 파일 변경으로 인한 경로 수정
